cri: move TestGenerateRuntimeOptions to the cri/config package
